refactor(delagent) use template
