+++ src/composer.json
@@ -35,6 +35,7 @@
             "Fossology\\Agent\\Copyright\\UI\\": "copyright/ui",
             "Fossology\\DeciderJob\\UI\\": "deciderjob/ui",
             "Fossology\\Decider\\UI\\": "decider/ui",
+            "Fossology\\DelAgent\\UI\\": "delagent/ui",
             "Fossology\\UI\\Page\\": "www/ui/page"
         }
     },
