+++ src/delagent/ui/DeleteResponse.php
@@ -13,8 +13,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  ***************************************************************/
 
-
-namespace delagent\ui;
+namespace Fossology\DelAgent\UI;
 
 class DeleteResponse
 {
