+++ src/delagent/ui/Page/AdminUploadDelete.php
@@ -1,0 +1,192 @@
+<?php
+/***********************************************************
+ Copyright (C) 2008-2013 Hewlett-Packard Development Company, L.P.
+ Copyright (C) 2015-2017 Siemens AG
+
+ This program is free software; you can redistribute it and/or
+ modify it under the terms of the GNU General Public License
+ version 2 as published by the Free Software Foundation.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License along
+ with this program; if not, write to the Free Software Foundation, Inc.,
+ 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+***********************************************************/
+namespace Fossology\DelAgent\UI\Page;
+
+use Fossology\Lib\Auth\Auth;
+use Fossology\Lib\Dao\UploadDao;
+use Fossology\Lib\Plugin\DefaultPlugin;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Fossology\DelAgent\UI\DeleteMessages;
+use Fossology\DelAgent\UI\DeleteResponse;
+
+class AdminUploadDelete extends DefaultPlugin
+{
+  const NAME = "admin_upload_delete";
+  /** @var UploadDao */
+  private $uploadDao;
+  function __construct()
+  {
+    parent::__construct(self::NAME, array(
+        self::TITLE => _("Delete Uploaded File"),
+        self::MENU_LIST => "Organize::Uploads::Delete Uploaded File",
+        self::PERMISSION => Auth::PERM_ADMIN,
+        self::REQUIRES_LOGIN => TRUE
+    ));
+    
+    global $container;
+    $this->uploadDao = $container->get('dao.upload');
+  }
+
+  
+  /**
+   * @param int $uploadpk - the upload(upload_id) you want to delete
+   * \return NULL on success, string on failure.
+   */
+  private function delete($uploadpk) 
+  {
+    /* Prepare the job: job "Delete" */
+    $user_pk = Auth::getUserId();
+    $group_pk = Auth::getGroupId();
+    $jobpk = JobAddJob($user_pk, $group_pk, "Delete", $uploadpk);
+    if (empty($jobpk) || ($jobpk < 0)) {
+      return _("Failed to create job record");
+    }
+    /* Add job: job "Delete" has jobqueue item "delagent" */
+    $jqargs = "DELETE UPLOAD $uploadpk";
+    $jobqueuepk = JobQueueAdd($jobpk, "delagent", $jqargs, NULL, NULL);
+    if (empty($jobqueuepk)) {
+      return _("Failed to place delete in job queue");
+    }
+
+    /* Tell the scheduler to check the queue. */
+    $success  = fo_communicate_with_scheduler("database", $output, $error_msg);
+    if (!$success) {
+      $error_msg = _("Is the scheduler running? Your jobs have been added to job queue.");
+      $URL = Traceback_uri() . "?mod=showjobs&upload=$uploadpk ";
+      $LinkText = _("View Jobs");
+      return "$error_msg <a href=\"$URL\">$LinkText</a>";
+    }
+    return NULL;
+  }
+  
+    /**
+   * @param Request $request
+   * @return Response
+   */
+  protected function handle(Request $request)
+  {
+    $vars = array();
+    
+    $uploadpks = $request->get('uploads');
+    if (!empty($uploadpks))
+    {
+      $vars['message'] = $this->initDeletion($uploadpks);
+    }
+
+    $vars['uploadScript'] = ActiveHTTPscript("Uploads");
+    $vars['tracbackUri'] = Traceback_uri();
+    $root_folder_pk = GetUserRootFolder();
+    $vars['rootFolderListOptions'] = FolderListOption($root_folder_pk, 0);
+    
+    $uploadList = array();
+    $folderList = FolderListUploads_perm($root_folder_pk, Auth::PERM_WRITE);
+    foreach($folderList as $L) {
+      $desc = $L['name'];
+      if (!empty($L['upload_desc'])) {
+        $desc .= " (" . $L['upload_desc'] . ")";
+      }
+      if (!empty($L['upload_ts'])) {
+        $desc .= " :: " . substr($L['upload_ts'], 0, 19);
+      }
+      $uploadList[$L['upload_pk']] = $desc;
+    }
+    $vars['uploadList'] = $uploadList;
+    
+    return $this->render('admin_upload_delete.html.twig', $this->mergeWithDefault($vars));    
+  }
+  
+  
+    /**
+   * @param int[] $uploadpks
+   * @brief starts deletion and handles error messages
+   * @return string
+   */
+  private function initDeletion($uploadpks)
+  {
+    if(sizeof($uploadpks) <= 0)
+    {
+      return _("No uploads selected");
+    }
+
+    $errorMessages = [];
+    $deleteResponse = NULL;
+    foreach($uploadpks as $uploadPk)
+    {
+      $deleteResponse = $this->TryToDelete(intval($uploadPk));
+
+      if($deleteResponse->getDeleteMessageCode() != DeleteMessages::SUCCESS)
+      {
+        $errorMessages[] = $deleteResponse;
+      }
+    }
+
+    if(sizeof($uploadpks) == 1)
+    {
+      return $deleteResponse->getDeleteMessageString().$deleteResponse->getAdditionalMessage();
+    }
+
+    $displayMessage = "";
+    $countErrorMessages = array_count_values($errorMessages);
+    if (in_array(DeleteMessages::SCHEDULING_FAILED, $errorMessages)) {
+      $displayMessage .= "<br/>Scheduling failed for " .
+              $countErrorMessages[DeleteMessages::SCHEDULING_FAILED] . " uploads<br/>";
+    }
+
+    if (in_array(DeleteMessages::NO_PERMISSION, $errorMessages)) {
+      $displayMessage .= "No permission to delete " .
+              $countErrorMessages[DeleteMessages::NO_PERMISSION] . " uploads<br/>";
+    }
+
+    $displayMessage .= "Deletion of " .
+            (sizeof($uploadpks) - sizeof($errorMessages)) . " projects queued";
+    return DisplayMessage($displayMessage);
+  }
+  
+  /**
+   * \brief Given a folder_pk, try to add a job after checking permissions.
+   * \param $uploadpk - the upload(upload_id) you want to delete
+   *
+   * \return string with the message.
+   */
+  private function TryToDelete($uploadpk)
+  {
+    if(!$this->uploadDao->isEditable($uploadpk, Auth::getGroupId())){
+      $returnMessage = DeleteMessages::NO_PERMISSION;
+      return new DeleteResponse($returnMessage);
+    }
+
+    $rc = $this->delete(intval($uploadpk));
+
+    if (! empty($rc)) {
+      $returnMessage = DeleteMessages::SCHEDULING_FAILED;
+      return new DeleteResponse($returnMessage);
+    }
+
+    /* Need to refresh the screen */
+    $URL = Traceback_uri() . "?mod=showjobs&upload=$uploadpk ";
+    $LinkText = _("View Jobs");
+    $returnMessage = DeleteMessages::SUCCESS;
+    return new DeleteResponse($returnMessage,
+      " <a href=$URL>$LinkText</a>");
+  }
+}
+
+
+register_plugin(new AdminUploadDelete());
