+++ src/delagent/ui/services.php
@@ -1,0 +1,3 @@
+<?php
+$loader = $GLOBALS['container']->get('twig.loader');
+$loader->addPath(dirname(__FILE__).'/template');
