+++ src/www/ui/Makefile
@@ -24,8 +24,9 @@ CONFDIR = $(DESTDIR)$(SYSCONFDIR)
 DIRS = css images locale scripts ajaxPHP
 
 UIFILES = `find . -type f | grep -v svn |grep -v tests | grep -E "(php|dat|dtd|js|html|twig)$$"`
-OBSOLETEFILES = admin-tag-ns.php admin-change-owner.php admin-tag-ns-perm.php \
-                group-manage-self.php group-manage.php ajax-perms.php
+OBSOLETEFILES = admin-tag-ns.php admin-change-owner.php admin-tag-ns-perm.php admin-folder-delete.php \
+                group-manage-self.php group-manage.php ajax-perms.php \
+                upload_permissions.php upload-srv-files.php upload-vcs.php
 
 OTHERFILES = `find . -type f | grep -v svn |grep -v tests | grep -E "(css|htc|gif|png|ico|htm)$$"`
